chore(dal): move schema migrations to the driver
- Move schema migration functions into the driver impl for ease of
  access to the driver cache and methods
- Use the cached identity func argument instead of refetching it
- Move core (i.e. main or primary) aws migrations into a new submodule
- Rename the driver to "BuiltinSchemaDriver" (same naming structure as
  BuiltinSchemaHelpers)
- Ensure the driver is a private struct
- Use FuncId instead of name for BuiltinMissingFuncArgument error (and
  fix the output along the way)
