+++ lib/dal/src/builtins/schema/aws/data/aws_instance_types.json
