+++ lib/dal/src/builtins/schema/aws/data/aws_regions.json
